requirements.txt
django
djangorestframework
psycopg2-binary
pillow
python-decouple
dj-config-url